requirements.txt
openai
anthropic
google-generativeai
python-dotenv
grpcio==1.60.1